fix circ deps
