+++ packages/paperclip-cli-utils/README.md
@@ -1,0 +1,4 @@
+CLI tool for compiling paperclip templates.
+
+Installation: `npm install paperclip-cli --save-dev`
+Documentation: http://paperclip.dev/docs/usage-cli
