+++ packages/paperclip-cli-utils/bin/paperclip
@@ -1,0 +1,71 @@
+#!/usr/bin/env node
+const path = require("path");
+const resolve = require("resolve");
+const chokidar = require("chokidar");
+const glob = require("glob");
+const {build, init, devStart} = require("../lib/cli");
+const fs = require("fs");
+
+// const argv = require("yargs")
+// .option("compiler", {
+//   describe: "Language compiler target",
+// })
+// .option("write", {
+//   describe: "Write compiled file",
+// })
+// .option("config", {
+//   describe: "Config file",
+// })
+// .option("drop-pc-extension", {
+//   describe: "Drops PC extension from compiled files",
+// })
+// .option("watch", {
+// })
+// .example("$0", "Runs paperclip using paperclip.config.json in cwd")
+// .example("$0 **/*.pc --compiler=paperclip-compiler-react", "Compiles PC files to React")
+// .example("$0 path/to/file.pc --compiler=paperclip-compiler-react --definition", "Generates a Typed Definition file")
+// .example("$0 --definition --write --watch", "Writes typed definition files & re-generates them when they change")
+// .argv;
+
+const argv = require("yargs")
+.command("build [file-pattern]", "Compiles paperclip files to target language", yargs => {
+  yargs
+  .option("watch", desc("watch for changes"))
+  .option("drop-pc-extension", desc("Drops PC extension from compiled files"))
+  .option("write", desc("Write compiled file"))
+  .option("compiler", desc("Language compiler target (default is from paperclip.config.json if found)"))
+  .option("definition", desc("Write typed definition type"))
+}, argv => {
+  build({
+    config: argv.config,
+    write: argv.write,
+    watch: argv.watch,
+    definition: argv.definition,
+    dropPcExtension: argv.dropPcExtension,
+    compilerName: argv.compiler,
+    filePatterns: argv.filePattern
+  })
+})
+.command("init", "Initialize new paperclip project", yargs => {
+}, argv => {
+  init();
+})
+.command("dev", "Starts the dev server", yargs => {
+  yargs
+  .option("port", desc("HTTP port to listen on"))
+  .default("port", 4000)
+}, argv => {
+  devStart({
+    port: argv.port,
+    cwd: process.cwd()
+  });
+})
+.argv;
+
+
+
+function desc(description) {
+  return {
+    description
+  }
+}
