+++ packages/paperclip-cli-utils/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "outDir": "lib",
+    "target": "es5",
+    "module": "commonjs",
+    "esModuleInterop": true,
+    "jsx": "react"
+  },
+  "exclude": [
+    "node_modules",
+    "*-test",
+    "lib",
+    "index.d.ts",
+    "test",
+    "examples"
+  ],
+  "filesGlob": ["./src/**/*.ts"]
+}
