+++ packages/paperclip-cli/src/cli.ts
@@ -1,4 +1,4 @@
 export * from "./init";
 export * from "./build";
-export * from "./pretty-message";
+export * from "paperclip-cli-utils";
 export * from "./dev";
