+++ packages/paperclip-cli/src/dev.ts
@@ -1,8 +1,6 @@
 import * as path from "path";
 import * as fs from "fs";
-import { startServer } from "paperclip-visual-editor";
 import { PC_CONFIG_FILE_NAME, createEngine, PaperclipConfig } from "paperclip";
-import { exec } from "child_process";
 
 export type ServerOptions = {
   port?: number;
@@ -10,6 +8,18 @@ export type ServerOptions = {
 };
 
 export const devStart = async ({ port, cwd }: ServerOptions) => {
+  let module;
+
+  try {
+    module = require("paperclip-visual-editor");
+  } catch (e) {
+    console.error(
+      `you'll need to install the visual editor: "npm install paperclip-visual-editor"`
+    );
+  }
+
+  const { startServer } = module;
+
   const config: PaperclipConfig = JSON.parse(
     fs.readFileSync(path.join(cwd, PC_CONFIG_FILE_NAME), "utf8")
   ) as PaperclipConfig;
