+++ packages/paperclip-visual-editor/package.json
@@ -30,6 +30,7 @@
     "@types/lodash": "^4.14.159",
     "@types/node": "^14.0.27",
     "@types/react": "^16.9.46",
+    "paperclip-cli": "^14.5.3",
     "@types/react-dom": "^16.9.8",
     "@types/redux": "^3.6.0",
     "@types/redux-saga": "^0.10.5",
@@ -42,7 +43,6 @@
     "file-loader": "^6.0.0",
     "html-webpack-plugin": "^4.3.0",
     "lru-cache": "^6.0.0",
-    "paperclip-cli": "^14.5.3",
     "paperclip-compiler-react": "^14.5.3",
     "paperclip-loader": "^14.5.3",
     "style-loader": "^1.2.1",
